// frontend/src/types/ProductType.ts
export type ProductType = {
  name: string;
  images: Array<string>;
  size?: string;
  comfort?: string;
  best_for?: string;
  mattress_type?: string;
  cooling_technology?: string;
  motion_separation?: string;
  pressure_relief?: string;
  support?: string;
  adjustable_base_friendly?: string;
  breathable?: string;

  num_reviews?: Number;
  current_price?: Number;
  rating?: Number;
  original_price?: Number;
  mattress_in_a_box?: string;
  height?: Number;
};